@@ -1,58 +1,106 @@
 #ifndef DIJKSTRA_HPP
 #define DIJKSTRA_HPP
 
-#include <algorithm>
-#include <queue>
+#include <boost/heap/fibonacci_heap.hpp>
+#include <functional>
+#include <ranges>
+#include <type_traits>
+#include <vector>
 
 #include "typedefs.hpp"
 
-Point nextOrigin(const Array2D<double> &dist, const Array2D<uint8_t> &visited);
-std::vector<Point> compilePath(const Array2D<Point> &prev, const Point &dest);
+namespace detail {
 
-struct Priority {
-    Point p;
+struct Node {
+    Point positon;
     double priority;
+};
 
-    [[nodiscard]] friend constexpr auto operator<(
-        const Priority &lhs, const Priority &rhs) noexcept {
-        return lhs.priority > rhs.priority;
+struct CompareNodes {
+    bool operator()(const Node &l, const Node &r) const {
+        return l.priority > r.priority;
     }
 };
 
-template <class T, class Container = std::vector<T>,
-          class Compare = std::less<typename Container::value_type>>
-class MyQueue : public std::priority_queue<T, Container, Compare> {
-public:
-    typedef typename std::priority_queue<
-        T, Container, Compare>::container_type::const_iterator const_iterator;
+class Queue {
+    using Heap =
+        boost::heap::fibonacci_heap<Node, boost::heap::compare<CompareNodes>>;
 
-    const_iterator cbegin() const noexcept { return this->c.cbegin(); }
-    const_iterator cend() const noexcept { return this->c.cend(); }
+    std::vector<std::pair<Point, Heap::handle_type>> handles;
+    Heap heap;
+
+public:
+    Node extract() {
+        const auto top = heap.top();
+        handles.erase(std::remove_if(handles.begin(), handles.end(),
+                                     [&top](const auto &h) {
+                                         return h.first == top.positon;
+                                     }),
+                      handles.end());
+        heap.pop();
+        return top;
+    }
+
+    void add(const Point &p, double priority) {
+        const auto handle = heap.push({ p, priority });
+        handles.push_back({ p, handle });
+    }
+
+    void decrease(const Point &p, double priority) {
+        const auto handle =
+            std::find_if(handles.begin(), handles.end(),
+                         [&p](const auto &h) { return h.first == p; });
+        heap.decrease(handle->second, { p, priority });
+    }
+
+    bool empty() const { return heap.empty(); }
+
+    bool contains(const Point &p) const {
+        return std::find_if(handles.begin(), handles.end(),
+                            [&p](const auto &h) { return h.first == p; }) !=
+               handles.end();
+    }
 };
 
-template <typename T>
-std::vector<Point> dijkstra(const T &graph, const Point &source,
+template <class Enum>
+constexpr std::underlying_type_t<Enum> to_underlying(Enum e) noexcept {
+    return static_cast<std::underlying_type_t<Enum>>(e);
+}
+
+constexpr auto enum_range = [](auto front, auto back) {
+    return std::views::iota(to_underlying(front), to_underlying(back) + 1) |
+           std::views::transform([](auto e) { return decltype(front)(e); });
+};
+
+}  // namespace detail
+
+std::vector<Point> compilePath(const Array2D<Point> &prev, const Point &dest);
+
+template <typename Graph>
+std::vector<Point> dijkstra(const Graph &graph, const Point &source,
                             const Point &dest) {
     Array2D<double> dist{ graph.w(), graph.h(), INF };
     Array2D<Point> prev{ graph.w(), graph.h(), { -1, -1 } };
     dist[source] = 0;
 
-    MyQueue<Priority> Q;
-    Q.push({ source, dist[source] });
+    detail::Queue queue;
+    queue.add(source, dist[source]);
 
-    while (!Q.empty()) {
-        const auto u = Q.top();
-        Q.pop();
-        if (u.priority == dist[u.p]) {
-            for (Dir d{}; d != Dir::COUNT; ++d) {
-                const auto v = u.p + d;
-                if (v.x >= 0 && v.y >= 0 && v.x < graph.w() &&
-                    v.y < graph.h()) {
-                    const auto alt = dist[u.p] + graph.distance(u.p, d);
-                    if (alt < dist[v] && dist[u.p] != INF) {
-                        dist[v] = alt;
-                        prev[v] = u.p;
-                        Q.push({ v, alt });
+    while (!queue.empty()) {
+        const auto u = queue.extract();
+        for (const auto dir :
+             detail::enum_range(Dir::RIGHT, Dir::BOTTOMRIGHT)) {
+            const auto v = u.positon + dir;
+            if (v.x >= 0 && v.y >= 0 && v.x < graph.w() && v.y < graph.h()) {
+                const auto alt =
+                    dist[u.positon] + graph.distance(u.positon, dir);
+                if (alt < dist[v] && dist[u.positon] != INF) {
+                    dist[v] = alt;
+                    prev[v] = u.positon;
+                    if (queue.contains(v)) {
+                        queue.decrease(v, alt);
+                    } else {
+                        queue.add(v, alt);
                     }
                 }
             }
